Better condition

+++ heroku/0.0.1/home/scripts/entrypoint.sh
@@ -2,15 +2,13 @@
 
 mkdir -p $HOME/.local/share/heroku
 
-if [ -f $HOME/config/config.json ]; then
-  if [ -f $HOME/config/.netrc ]; then
-    cp $HOME/config/config.json $HOME/.local/share/heroku/
-    cp $HOME/config/.netrc $HOME
-  fi
+if [ -f $HOME/config/config.json ] && [ -f $HOME/config/.netrc ]; then
+  cp $HOME/config/config.json $HOME/.local/share/heroku/
+  cp $HOME/config/.netrc $HOME
 fi
 
 ACTION=$1
-  shift 1
+shift 1
 
 if [ -f $HOME/scripts/$ACTION.sh ]; then
   $HOME/scripts/$ACTION.sh $*
